+++ src/Service/CodeDeploy/src/Exception/InvalidECSServiceException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\CodeDeploy\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * The Amazon ECS service identifier is not valid.
+ */
+final class InvalidECSServiceException extends ClientException
+{
+}