+++ src/Service/CodeDeploy/src/Exception/InvalidInputException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\CodeDeploy\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * The input was specified in an invalid format.
+ */
+final class InvalidInputException extends ClientException
+{
+}